+++ PharmaceuticalWarehouseManagementSystem/PharmaceuticalWarehouseManagementSystem.UI/Areas/Admin/Controllers/OrderController.cs
@@ -18,19 +18,19 @@ namespace PharmaceuticalWarehouseManagementSystem.UI.Areas.Admin.Controllers
     [Authorize(Roles = "Admin")]
     public class OrderController : Controller
     {
-        private IOrderRepository _repository;
-        private ProjectContext _context;
-        private ILogger<OrderController> _Ilogger;
+        private readonly IOrderRepository _repository;
+        private readonly ProjectContext _context;
+        private readonly ILogger<OrderController> _logger;
 
         public OrderController(IOrderRepository repository, ProjectContext context,ILogger<OrderController> logger)
         {
             this._repository = repository;
             this._context = context;
-            this._Ilogger = logger;
+            this._logger = logger;
         }
         public IActionResult List()
         {
-            _Ilogger.LogInformation("Orders Listed "+DateTime.Now.ToString());
+            _logger.LogInformation("Orders Listed "+DateTime.Now.ToString());
             return View(_repository.GetActive());
         }
 
@@ -59,20 +59,20 @@ namespace PharmaceuticalWarehouseManagementSystem.UI.Areas.Admin.Controllers
                 if (result == true)
                 {
                     _repository.Save();
-                    _Ilogger.LogInformation("Order Added "+" "+DateTime.Now.ToString());
+                    _logger.LogInformation("Order Added "+" "+DateTime.Now.ToString());
                     return RedirectToAction("List");
                 }
                 else
                 {
                     TempData["Message"] = $"Kayıt işlemi sırasında bir hata oluştu. Lütfen tüm alanları kontrol edip tekrar deneyin..!";
-                    _Ilogger.LogError("Order add failed "+DateTime.Now.ToString());
+                    _logger.LogError("Order add failed "+DateTime.Now.ToString());
                     return View(item);
                 }
             }
             else
             {
                 TempData["Message"] = $"Kayıt işlemi sırasında bir hata oluştu. Lütfen tüm alanları kontrol edip tekrar deneyin..!";
-                _Ilogger.LogCritical("Order add failed");
+                _logger.LogCritical("Order add failed");
                 return View(item);
             }
         }
@@ -107,29 +107,37 @@ namespace PharmaceuticalWarehouseManagementSystem.UI.Areas.Admin.Controllers
                 if (result)
                 {
                     _repository.Save();
-                    _Ilogger.LogInformation("Order "+item.ID+" "+"Edited "+DateTime.Now.ToString());
+                    _logger.LogInformation("Order "+item.ID+" "+"Edited "+DateTime.Now.ToString());
                     return RedirectToAction("List");
                 }
                 else
                 {
                     TempData["Message"] = $"Güncelleme işlemi sırasında bir hata oluştu. Lütfen tekrar deneyin..!";
-                    _Ilogger.LogError("Order Edit Error" + DateTime.Now.ToString());
+                    _logger.LogError("Order Edit Error" + DateTime.Now.ToString());
                     return View(updated);
                 }
             }
             else
             {
                 TempData["Message"] = $"Güncelleme işlemi sırasında bir hata oluştu. Lütfen tekrar deneyin..!";
-                _Ilogger.LogCritical("Order Edit Error");
+                _logger.LogCritical("Order Edit Error");
                 return View();
             }
         }
 
         public IActionResult Delete(Guid id)
         {
-            _repository.Remove(_repository.GetById(id));
-            _Ilogger.LogInformation("Order "+id+" "+"Deleted "+DateTime.Now.ToString());
-            return RedirectToAction("List");
+            if (ModelState.IsValid)
+            {
+                _logger.LogInformation("Order Deleted"+" "+ id+" "+DateTime.Now.ToString());
+                _repository.Remove(_repository.GetById(id));
+                return RedirectToAction("List");
+            }
+            else
+            {
+                _logger.LogError("Order Delete Action Failed"+" "+DateTime.Now.ToString());
+                return BadRequest();
+            }
         }
